+++ src/main/java/com/github/castorm/kafka/connect/http/record/SimpleKvSourceRecordMapper.java
@@ -23,40 +23,46 @@ package com.github.castorm.kafka.connect.http.record;
 import com.github.castorm.kafka.connect.http.model.Offset;
 import com.github.castorm.kafka.connect.http.record.model.KvRecord;
 import com.github.castorm.kafka.connect.http.record.spi.KvSourceRecordMapper;
+import lombok.RequiredArgsConstructor;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.SchemaBuilder;
 import org.apache.kafka.connect.data.Struct;
 import org.apache.kafka.connect.source.SourceRecord;
 
 import java.util.Map;
+import java.util.function.Function;
 
 import static java.util.Collections.emptyMap;
 import static org.apache.kafka.connect.data.SchemaBuilder.string;
 
+@RequiredArgsConstructor
 public class SimpleKvSourceRecordMapper implements KvSourceRecordMapper {
 
-    private static final String SCHEMA_KEY_KEY = "key";
-    private static final Schema SCHEMA_KEY = SchemaBuilder.struct()
-            .name("com.github.castorm.kafka.connect.http.Key")
-            .doc("Message Key")
-            .field(SCHEMA_KEY_KEY, string().optional().doc("HTTP Item identifier").build())
-            .build();
+    private static Map<String, ?> sourcePartition = emptyMap();
 
-    private static final String SCHEMA_VALUE_BODY = "body";
-    private static final Schema SCHEMA_VALUE = SchemaBuilder.struct()
-            .name("com.github.castorm.kafka.connect.http.Value")
-            .doc("Message Value")
-            .field(SCHEMA_VALUE_BODY, string().doc("HTTP Item response body").build())
-            .build();
+    private final Function<Map<String, ?>, SimpleKvSourceRecordMapperConfig> configFactory;
 
-    private Map<String, ?> sourcePartition = emptyMap();
+    private SimpleKvSourceRecordMapperConfig config;
 
-    private String topic;
+    private Schema keySchema;
+
+    private Schema valueSchema;
+
+    public SimpleKvSourceRecordMapper() {
+        this(SimpleKvSourceRecordMapperConfig::new);
+    }
 
     @Override
-    public void configure(Map<String, ?> configs) {
-        SimpleKvSourceRecordMapperConfig config = new SimpleKvSourceRecordMapperConfig(configs);
-        topic = config.getTopic();
+    public void configure(Map<String, ?> settings) {
+        config = configFactory.apply(settings);
+        keySchema = SchemaBuilder.struct()
+                .name("com.github.castorm.kafka.connect.http.Key").doc("Message Key")
+                .field(config.getKeyPropertyName(), string().optional().doc("HTTP Item key").build())
+                .build();
+        valueSchema = SchemaBuilder.struct()
+                .name("com.github.castorm.kafka.connect.http.Value").doc("Message Value")
+                .field(config.getValuePropertyName(), string().doc("HTTP Item value").build())
+                .build();
     }
 
     @Override
@@ -69,7 +75,7 @@ public class SimpleKvSourceRecordMapper implements KvSourceRecordMapper {
         return new SourceRecord(
                 sourcePartition,
                 offset.toMap(),
-                topic,
+                config.getTopic(),
                 null,
                 key.schema(),
                 key,
@@ -78,11 +84,11 @@ public class SimpleKvSourceRecordMapper implements KvSourceRecordMapper {
                 offset.getTimestamp().toEpochMilli());
     }
 
-    private static Struct keyStruct(String key) {
-        return new Struct(SCHEMA_KEY).put(SCHEMA_KEY_KEY, key);
+    private Struct keyStruct(String key) {
+        return new Struct(keySchema).put(config.getKeyPropertyName(), key);
     }
 
-    private static Struct valueStruct(String body) {
-        return new Struct(SCHEMA_VALUE).put(SCHEMA_VALUE_BODY, body);
+    private Struct valueStruct(String body) {
+        return new Struct(valueSchema).put(config.getValuePropertyName(), body);
     }
 }