+++ src/test/java/com/github/castorm/kafka/connect/http/record/SimpleKvSourceRecordMapperConfigTest.java
@@ -24,6 +24,10 @@ import com.google.common.collect.ImmutableMap;
 import org.apache.kafka.common.config.ConfigException;
 import org.junit.jupiter.api.Test;
 
+import java.util.HashMap;
+import java.util.Map;
+
+import static com.github.castorm.kafka.connect.http.record.SimpleKvSourceRecordMapperConfigTest.Fixture.minimumConfig;
 import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
@@ -37,6 +41,36 @@ class SimpleKvSourceRecordMapperConfigTest {
 
     @Test
     void whenKafkaTopic_thenInitialized() {
-        assertThat(new SimpleKvSourceRecordMapperConfig(ImmutableMap.of("kafka.topic", "test-topic")).getTopic()).isEqualTo("test-topic");
+        assertThat(minimumConfig(emptyMap()).getTopic()).isEqualTo("test-topic");
+    }
+
+    @Test
+    void whenMissingKeyProperty_thenDefault() {
+        assertThat(minimumConfig(emptyMap()).getKeyPropertyName()).isEqualTo("key");
+    }
+
+    @Test
+    void whenKeyProperty_thenInitialized() {
+        assertThat(minimumConfig(ImmutableMap.of("http.record.schema.key.property.name", "custom")).getKeyPropertyName()).isEqualTo("custom");
+    }
+
+    @Test
+    void whenMissingValueProperty_thenDefault() {
+        assertThat(minimumConfig(emptyMap()).getValuePropertyName()).isEqualTo("value");
+    }
+
+    @Test
+    void whenValueProperty_thenInitialized() {
+        assertThat(minimumConfig(ImmutableMap.of("http.record.schema.value.property.name", "custom")).getValuePropertyName()).isEqualTo("custom");
+    }
+
+    interface Fixture {
+
+        static SimpleKvSourceRecordMapperConfig minimumConfig(Map<String, String> customConfig) {
+            HashMap<String, String> finalConfig = new HashMap<>();
+            finalConfig.put("kafka.topic", "test-topic");
+            finalConfig.putAll(customConfig);
+            return new SimpleKvSourceRecordMapperConfig(finalConfig);
+        }
     }
 }