+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   `timestamp_iso` key in `Offset` renamed to plain `timestamp`
 -   `key` added to `Offset`
 -   `JacksonKvRecordHttpResponseParser` to generate a consistent `UUID` as key when key is missing
--   `AdaptableIntervalThrottler` to skip throttling on first request 
+-   `AdaptableIntervalThrottler` to skip throttling on first request
+-   `SimpleKvSourceRecordMapper` to support custom key/value property names 
 
 ### v0.7.0 (05/23/2020)
 -   `http.response.records.pointer` renamed to `http.response.list.pointer`